requirements.txt
qdrant-client
ollama
mcp
marker-pdf
anthropic
openai
pyyaml
httpx
streamlit
fastapi
uvicorn
fastembed
